Claude bsed refactoring of the base app service
