+++ register-api/app/services/invenio_rdm_service.py
@@ -17,20 +17,20 @@ import json
 
 class IvenioRDMService:
 
-    # Uncomment this block if you'd like way too much info on the http requests being made to RDM
+    # # Uncomment this block if you'd like way too much info on the http requests being made to RDM
 
-    try:
-        import http.client as http_client
-    except ImportError:
-        # Python 2
-        import httplib as http_client
-    http_client.HTTPConnection.debuglevel = 1
-    # You must initialize logging, otherwise you'll not see debug output.
-    logging.basicConfig()
-    logging.getLogger().setLevel(logging.DEBUG)
-    requests_log = logging.getLogger("requests.packages.urllib3")
-    requests_log.setLevel(logging.DEBUG)
-    requests_log.propagate = True
+    # try:
+    #     import http.client as http_client
+    # except ImportError:
+    #     # Python 2
+    #     import httplib as http_client
+    # http_client.HTTPConnection.debuglevel = 1
+    # # You must initialize logging, otherwise you'll not see debug output.
+    # logging.basicConfig()
+    # logging.getLogger().setLevel(logging.DEBUG)
+    # requests_log = logging.getLogger("requests.packages.urllib3")
+    # requests_log.setLevel(logging.DEBUG)
+    # requests_log.propagate = True
     
 
     def __init__(self, url: string, token:string):
